removed casting

+++ src/Sql/PostgreSQL/Functions/user_create.sql
@@ -1,4 +1,4 @@
-DROP FUNCTION IF EXISTS user_create(UUID,VARCHAR,VARCHAR,BOOLEAN,VARCHAR,VARCHAR,VARCHAR,VARCHAR,TEXT,VARCHAR,TEXT,TEXT,TIMESTAMPTZ,TEXT,TEXT,TEXT,BOOLEAN,TIMESTAMPTZ,TIMESTAMPTZ,BIGINT,INT,INT,VARCHAR,VARCHAR,VARCHAR,INT,INT,TIMESTAMPTZ,TIMESTAMPTZ);
+DROP FUNCTION IF EXISTS user_create(UUID,VARCHAR,VARCHAR,BOOLEAN,VARCHAR,VARCHAR,VARCHAR,VARCHAR,TEXT,VARCHAR,TEXT,TEXT,TIMESTAMPTZ,TEXT,TEXT,TEXT,BOOLEAN,TIMESTAMPTZ,TIMESTAMPTZ,BIGINT,SMALLINT,SMALLINT,VARCHAR,VARCHAR,VARCHAR,SMALLINT,INT,TIMESTAMPTZ,TIMESTAMPTZ);
 
 CREATE OR REPLACE FUNCTION user_create
 (
@@ -22,12 +22,12 @@ CREATE OR REPLACE FUNCTION user_create
     _premium_expiration_date            TIMESTAMPTZ,
     _renewal_reminder_date              TIMESTAMPTZ,
     _storage                            BIGINT,
-    _max_storage_gb                     INT,
-    _gateway                            INT,
+    _max_storage_gb                     SMALLINT,
+    _gateway                            SMALLINT,
     _gateway_customer_id                VARCHAR,
     _gateway_subscription_id            VARCHAR,
     _license_key                        VARCHAR,
-    _kdf                                INT,
+    _kdf                                SMALLINT,
     _kdf_iterations                     INT,
     _creation_date                      TIMESTAMPTZ,
     _revision_date                      TIMESTAMPTZ
@@ -91,12 +91,12 @@ BEGIN
         _premium_expiration_date,
         _renewal_reminder_date,
         _storage,
-        _max_storage_gb::SMALLINT,
-        _gateway::SMALLINT,
+        _max_storage_gb,
+        _gateway,
         _gateway_customer_id,
         _gateway_subscription_id,
         _license_key,
-        _kdf::SMALLINT,
+        _kdf,
         _kdf_iterations,
         _creation_date,
         _revision_date
